rhods_notebook_api_scale_test: update
