--- roles/rhods_notebook_api_scale_test/files/entrypoint/configmaps_envs.json
@@ -1,9 +1,0 @@
-{
-  "apiVersion": "v1",
-  "kind": "ConfigMap",
-  "metadata": {
-    "name": "jupyterhub-singleuser-profile-{user_name}-envs",
-    "namespace": "rhods-notebooks"
-  },
-  "data": {}
-}
