--- roles/rhods_notebook_api_scale_test/files/entrypoint/notebooks_terminate.json
@@ -1,7 +1,0 @@
-{
-  "metadata": {
-    "annotations": {
-      "kubeflow-resource-stopped": "stopped by ci-artifacts"
-    }
-  }
-}
