--- roles/rhods_notebook_api_scale_test/files/entrypoint/pvc_storage.json
@@ -1,19 +1,0 @@
-{
-  "apiVersion": "v1",
-  "kind": "PersistentVolumeClaim",
-  "metadata": {
-    "name": "jupyterhub-nb-{user_name}-pvc",
-    "namespace": "rhods-notebooks"
-  },
-  "spec": {
-    "accessModes": [
-      "ReadWriteOnce"
-    ],
-    "resources": {
-      "requests": {
-        "storage": "5Gi"
-      }
-    },
-    "volumeMode": "Filesystem"
-  }
-}
