--- roles/rhods_notebook_api_scale_test/files/entrypoint/role_notebook-view.json
@@ -1,24 +1,0 @@
-{
-    "apiVersion": "rbac.authorization.k8s.io/v1",
-    "kind": "Role",
-    "metadata": {
-        "name": "jupyter-nb-{user_name}-notebook-view",
-        "namespace": "rhods-notebooks"
-    },
-    "rules": [
-        {
-            "apiGroups": [
-                "kubeflow.org"
-            ],
-            "resourceNames": [
-                "jupyter-nb-{user_name}"
-            ],
-            "resources": [
-                "notebooks"
-            ],
-            "verbs": [
-                "get"
-            ]
-        }
-    ]
-}
