--- roles/rhods_notebook_api_scale_test/files/entrypoint/secrets_envs.json
@@ -1,10 +1,0 @@
-{
-  "apiVersion": "v1",
-  "kind": "Secret",
-  "metadata": {
-    "name": "jupyterhub-singleuser-profile-{user_name}-envs",
-    "namespace": "rhods-notebooks"
-  },
-  "stringData": {},
-  "type": "Opaque"
-}
