+++ roles/rhods_notebook_api_scale_test/vars/main/resources.yml
@@ -1,4 +1,3 @@
 rhods_notebook_api_scale_test_job: roles/rhods_notebook_api_scale_test/templates/api_scale_test_job.yaml
 rhods_notebook_api_scale_test_entrypoint_directory: roles/rhods_notebook_api_scale_test/files/entrypoint
 rhods_notebook_api_scale_test_s3_artifacts_exporter_sidecar: roles/rhods_notebook_api_scale_test/files/s3-artifacts-exporter.sh
-rhods_notebook_api_scale_test_timescale_schema: roles/rhods_notebook_api_scale_test/files/locust-plugins_timescale/timescale_schema.sql
